backend/flask-api/requirements.txt
Flask==2.2.2
Flask_JWT_Extended==4.4.4
flask_sqlalchemy==3.0.2
flask_swagger_ui==4.11.1
python-dotenv==0.21.0
python_bcrypt==0.3.2
SQLAlchemy==1.4.44
psycopg2==2.9.5
